requirements.txt
mcp>=0.1.0
sqlalchemy>=2.0.0
faker>=20.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
tqdm>=4.65.0